% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amphibia_dates}
\alias{amphibia_dates}
\title{Example amphibian observations with dates}
\format{
A \code{data.frame} containing 224484 example observations described by 4 variables:
\itemize{
\item \code{species}: scientific name of observed species.
\item \code{code}: abbreviation of the scientific name.
\item \code{year}: year of observation.
\item \code{date}: date of observation.
}
}
\source{
A reduced subset prepared for package examples.
}
\usage{
amphibia_dates
}
\description{
A reduced set of amphibian observations with dates but no locations.
}
\keyword{datasets}
