% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aves_large}
\alias{aves_large}
\title{Large (species-wise) example bird-observation dataset}
\format{
A \code{data.frame} containing 132647 example observations from 200 species
described by 7 variables:
\itemize{
\item \code{species}: pseudonimised species.
\item \code{coordX}: approximate X-coordinate of observation (EPSG: 3059).
\item \code{coordY}: approximate Y-coordinate of observation (EPSG: 3059).
\item \code{year}: the year of observation.
\item \code{date}: the date (\code{as.Date} object) of the observation. Contains missing values.
\item \code{date_time}: the \code{POSIXct} object of date-time. Contains missing values.
\item \code{id100}: identificator of 100 m grid cell for joining with \code{data(points100)}.
}
}
\source{
A reduced and pseudonimised subset prepared for package examples.
}
\usage{
aves_large
}
\description{
A reduced bird-observation dataset intended for examples that require
multiple observations, species, dates, or times. Contains 132647 observations
from 200 species.
}
\keyword{datasets}
