% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aves_small}
\alias{aves_small}
\title{Small (species-wise) example bird-observation dataset}
\format{
A \code{data.frame} containing 6359 example observations from 5 species
described by 6 variables:
\itemize{
\item \code{species}: scientific name of observed species.
\item \code{code}: abbreviation of the scientific name.
\item \code{year}: the year of observation.
\item \code{date}: the date (\code{as.Date} object) of the observation.
\item \code{date_time}: the \code{POSIXct} object of date-time. Contains missing values.
\item \code{id100}: identificator of 100 m grid cell for joining with \code{data(points100)}.
}
}
\source{
A reduced subset prepared for package examples.
}
\usage{
aves_small
}
\description{
A small bird-observation dataset intended for fast examples. Contains 6539 observations
from 5 species.
}
\keyword{datasets}
