% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bryophyta_one}
\alias{bryophyta_one}
\title{Example observations for one bryophyte taxon}
\format{
A \code{data.frame} containing 170 example observations of one bryophyte taxon
described by 6 variables:
\itemize{
\item \code{species}: pseudonimised species.
\item \code{coordX}: approximate X-coordinate of observation (EPSG: 3059).
\item \code{coordY}: approximate Y-coordinate of observation (EPSG: 3059).
\item \code{year}: the year of observation.
\item \code{date}: the date (\code{as.Date} object) of the observation. Contains missing values.
\item \code{id100}: identificator of 100 m grid cell for joining with \code{data(points100)}.
}
}
\source{
A reduced and pseudonimised subset prepared for package examples.
}
\usage{
bryophyta_one
}
\description{
A small subset containing 170 observations of one bryophyte taxon.
}
\keyword{datasets}
