% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egv_names}
\alias{egv_names}
\title{Names of example ecogeographical variables}
\format{
A \code{data.frame} describing 5 ecogeographical (environmental) variables
provided in \code{"inst/extdata/example_egv.tif"} with 3 fields:
\itemize{
\item \code{layername}: abbreviated layer name used for \code{terra::names()}.
\item \code{longname_english}: descriptive name of the variable in English.
\item \code{longname_latvia}: descriptive name of the variable in Latvian.
}
}
\usage{
egv_names
}
\description{
Descriptive names and abbreviations corresponding to the ecogeographical
variable layers supplied with the package examples.
}
\keyword{datasets}
