% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_background}
\alias{example_background}
\title{Example background locations}
\format{
A \code{data.frame} containing 40000 background location coordinates
with 2 fields:
\itemize{
\item \code{x}: approximate X-coordinate (EPSG: 3059).
\item \code{y}: approximate Y-coordinate (EPSG: 3059).
}
}
\source{
Prepared for package examples and practically used in species
distribution modelling.
}
\usage{
example_background
}
\description{
Spatial background locations used in fold-selection and model-evaluation
examples.
}
\keyword{datasets}
