% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egv_file.R
\name{example_egv}
\alias{example_egv}
\title{Example ecogeographical-variable raster}
\format{
A GeoTIFF file containing a multilayer raster in the coordinate
reference system used by the package's spatial examples. Layers are
described in \code{data(egv_names)}
}
\source{
A reduced subset of ecogeographical variables prepared for
\code{sdmhelpers} examples.
}
\description{
A small multilayer raster included with \code{sdmhelpers} for examples and
demonstrations of spatial workflows.
}
\details{
The raster is stored as a GeoTIFF file under the package's \code{extdata}
directory. It is not loaded automatically as an R data object.

Locate the installed file with:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{egv_path <- system.file(
  "extdata",
  "example_egv.tif",
  package = "sdmhelpers"
)
}\if{html}{\out{</div>}}

Read it as a \code{terra::SpatRaster} with:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{example_egv <- terra::rast(egv_path)
}\if{html}{\out{</div>}}
}
\examples{
egv_path <- system.file("extdata",
"example_egv.tif",
package = "sdmhelpers")

if (nzchar(egv_path) && requireNamespace("terra", quietly = TRUE)) {
  example_egv <- terra::rast(egv_path)
  example_egv
}

}
