% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_grid}
\alias{example_grid}
\title{Example raster-grid geometry}
\format{
A \code{SpatRaster} object created with \code{\link[terra:rast]{terra::rast()}}.
}
\source{
Derived from a reduced section of the Latvian 100-m raster grid.
}
\usage{
example_grid
}
\description{
A raster-grid object defining the spatial geometry, pixel alignment and CRS
used by the package's raster examples.
}
\details{
File contains only presence cells coded with value \code{1}.
}
\keyword{datasets}
