% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_independent_blocks}
\alias{example_independent_blocks}
\title{Example independent spatial blocks}
\format{
A \code{sf} object containing independent spatial-blocks as \code{POLYGON} and
their identifiers.
}
\source{
A reduced subset prepared for package examples from actual species
distribution modelling.
}
\usage{
example_independent_blocks
}
\description{
Fold or block identifiers used to demonstrate independent-data selection.
}
\keyword{datasets}
