% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_presences}
\alias{example_presences}
\title{Example species-presence locations}
\format{
A \code{data.frame} containing 2778 species presence location coordinates
with 2 fields:
\itemize{
\item \code{x}: approximate X-coordinate (EPSG: 3059).
\item \code{y}: approximate Y-coordinate (EPSG: 3059).
}
}
\source{
Prepared for package examples and practically used in species
distribution modelling.
}
\usage{
example_presences
}
\description{
Spatial presence records used in spatial KDE, fold-selection, and model
evaluation examples.
}
\keyword{datasets}
