% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habitat_inventories}
\alias{habitat_inventories}
\title{Example habitat inventories}
\format{
A data frame or spatial object containing inventory dates covering
127700 \code{data(example_raster)} centroids and 6 associated attributes:
\itemize{
\item \code{habitat_type}: character vector of bvroad habitat class:
\item \code{dune};
\item \code{forest};
\item \code{freshwater};
\item \code{grassland};
\item \code{heather};
\item \code{mire}.
\item \code{coordX}: X-coordinate of \code{data(example_raster)} cell's centre, matching \code{data(points100)} object (EPSG: 3059).
\item \code{coordY}: Y-coordinate of \code{data(example_raster)} cell's centre, matching \code{data(points100)} object (EPSG: 3059).
\item \code{year}: year of habitat inventory.
\item \code{date}: date (\code{as.Date()}) of habitat inventory
\item \code{id100}: identificator of 100 m grid cell for joining with \code{data(points100)}.
}
}
\source{
A reduced and anonymised subset prepared for package examples.
}
\usage{
habitat_inventories
}
\description{
Habitat-inventory observations used to demonstrate seasonal weighting
against species-activity dates.
}
\keyword{datasets}
