% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points100}
\alias{points100}
\title{Example 100-m grid points}
\format{
An \code{sf} object containing point geometries and grid identifiers:
\itemize{
\item \code{id}: identificator of 100 m grid cell.
\item \code{yes}: indicator for location in terrestrial Latvia (all the cells contain value \code{1}).
\item \code{tks50}: identificator of topographic map's 50 km page.
\item \code{rinda300}: identificator of 300 m grid cell.
\item \code{ID1km}: identificator of 1 km grid cell.
\item \code{rinda500}: identificator of 500 m grid cell.
\item \code{geom}: \verb{sf geometry} field.
}
}
\source{
Derived from a reduced section of the Latvian 100-m analysis grid.
}
\usage{
points100
}
\description{
Point centres of a reduced 100-m grid used in spatial examples.
}
\keyword{datasets}
