% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmhelpers-package.R
\docType{package}
\name{sdmhelpers-package}
\alias{sdmhelpers}
\alias{sdmhelpers-package}
\title{sdmhelpers: Helper Functions for Species Distribution Modelling}
\description{
The \code{sdmhelpers} package provides utility functions for species distribution
modelling workflows, including:
}
\details{
\itemize{
\item phenological weighting from circular day-of-year kernel densities;
\item temporal overlap estimation on seasonal and daily cycles;
\item spatial-fold selection for independent testing;
\item threshold-based model evaluation using the true skill statistic;
\item fast raster-based spatial kernel density estimation.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aavotins/sdmhelpers}
  \item \url{https://aavotins.github.io/sdmhelpers/}
  \item Report bugs at \url{https://github.com/aavotins/sdmhelpers/issues}
}

}
\author{
\strong{Maintainer}: Andris Avotins \email{avotins.puces@gmail.com} (\href{https://orcid.org/0000-0001-8582-0822}{ORCID})

Authors:
\itemize{
  \item Andris Avotins \email{avotins.puces@gmail.com} (\href{https://orcid.org/0000-0001-8582-0822}{ORCID})
}

}
\keyword{internal}
